refactor: engine pattern package and add duration support
