+++ pkg/engine/operator/operator.go
@@ -26,42 +26,31 @@ const (
 	NotInRange Operator = "!-"
 )
 
-// ReferenceSign defines the operator for anchor reference
-const ReferenceSign Operator = "$()"
-
 // GetOperatorFromStringPattern parses opeartor from pattern
 func GetOperatorFromStringPattern(pattern string) Operator {
 	if len(pattern) < 2 {
 		return Equal
 	}
-
 	if pattern[:len(MoreEqual)] == string(MoreEqual) {
 		return MoreEqual
 	}
-
 	if pattern[:len(LessEqual)] == string(LessEqual) {
 		return LessEqual
 	}
-
 	if pattern[:len(More)] == string(More) {
 		return More
 	}
-
 	if pattern[:len(Less)] == string(Less) {
 		return Less
 	}
-
 	if pattern[:len(NotEqual)] == string(NotEqual) {
 		return NotEqual
 	}
-
 	if match, _ := regexp.Match(`^(\d+(\.\d+)?)([^-]*)!-(\d+(\.\d+)?)([^-]*)$`, []byte(pattern)); match {
 		return NotInRange
 	}
-
 	if match, _ := regexp.Match(`^(\d+(\.\d+)?)([^-]*)-(\d+(\.\d+)?)([^-]*)$`, []byte(pattern)); match {
 		return InRange
 	}
-
 	return Equal
 }
